Increased use of NodeIndexSeek to wider range of cases
* Equality predicates on expressions of map literals and variables of map type
* Index Nested Loop joins even when incoming horizon contained a query plan
