+++ community/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/UsingAcceptanceTest.scala
@@ -692,58 +692,4 @@ class UsingAcceptanceTest extends ExecutionEngineFunSuite with NewPlannerTestSup
         intercept[QueryExecutionException](db.execute(s"EXPLAIN $query")).getStatusCode should equal("Neo.DatabaseError.Statement.ExecutionFailed")
     }
   }
-
-  case class includeOnlyOneHashJoinOn(nodeVariable: String) extends Matcher[InternalPlanDescription] {
-
-    private val hashJoinStr = classOf[NodeHashJoin].getSimpleName
-
-    override def apply(result: InternalPlanDescription): MatchResult = {
-      val hashJoins = result.flatten.filter { description =>
-        description.name == hashJoinStr && description.arguments.contains(KeyNames(Seq(nodeVariable)))
-      }
-      val numberOfHashJoins = hashJoins.length
-
-      MatchResult(numberOfHashJoins == 1, matchResultMsg(negated = false, result, numberOfHashJoins), matchResultMsg(negated = true, result, numberOfHashJoins))
-    }
-
-    private def matchResultMsg(negated: Boolean, result: InternalPlanDescription, numberOfHashJoins: Integer) =
-      s"$hashJoinStr on node '$nodeVariable' should exist only once in the plan description ${if (negated) "" else s", but it occurred $numberOfHashJoins times"}\n $result"
-  }
-
-  case class includeOnlyOne[T](operator: Class[T], withVariable: String = "") extends includeOnly(operator, withVariable) {
-    override def verifyOccurences(actualOccurences: Int) =
-      actualOccurences == 1
-
-    override def matchResultMsg(negated: Boolean, result: InternalPlanDescription, numberOfOperatorOccurences: Integer) =
-      s"$joinStr on node '$withVariable' should occur only once in the plan description${if (negated) "" else s", but it occurred $numberOfOperatorOccurences times"}\n $result"
-  }
-
-  case class includeAtLeastOne[T](operator: Class[T], withVariable: String = "") extends includeOnly(operator, withVariable) {
-    override def verifyOccurences(actualOccurences: Int) =
-      actualOccurences >= 1
-
-    override def matchResultMsg(negated: Boolean, result: InternalPlanDescription, numberOfOperatorOccurences: Integer) =
-      s"$joinStr on node '$withVariable' should occur at least once in the plan description${if (negated) "" else s", but it was not found\n $result"}"
-  }
-
-  abstract class includeOnly[T](operator: Class[T], withVariable: String = "") extends Matcher[InternalPlanDescription] {
-    protected val joinStr = operator.getSimpleName
-
-    def verifyOccurences(actualOccurences: Int): Boolean
-
-    def matchResultMsg(negated: Boolean, result: InternalPlanDescription, numberOfOperatorOccurences: Integer): String
-
-    override def apply(result: InternalPlanDescription): MatchResult = {
-      val operatorOccurrences = result.flatten.filter { description =>
-        val nameCondition = description.name == joinStr
-        val variableCondition = withVariable == "" || description.variables.contains(withVariable)
-        nameCondition && variableCondition
-      }
-      val numberOfOperatorOccurrences = operatorOccurrences.length
-      val matches = verifyOccurences(numberOfOperatorOccurrences)
-
-      MatchResult(matches, matchResultMsg(negated = false, result, numberOfOperatorOccurrences), matchResultMsg(negated = true, result, numberOfOperatorOccurrences))
-    }
-  }
-
 }
