+++ community/cypher/frontend-3.1/src/main/scala/org/neo4j/cypher/internal/frontend/v3_1/ast/Clause.scala
@@ -180,7 +180,7 @@ case class Match(optional: Boolean, pattern: Pattern, hints: Seq[UsingHint], whe
    */
   private def applicable(other: Expression) = {
     other match {
-      case _: Property => false
+//      case _: Property => false
       case f: FunctionInvocation => f.function != functions.Id
       case _ => true
     }
